requirements.txt
geopy==2.3.0
gspread==3.7.0
numpy==1.24.1
oauth2client==4.1.3
pandas==1.5.3
Pillow>=10.1
pytesseract==0.3.10
opencv-python==4.7.0.68